+++ alert/types.go
@@ -71,6 +71,9 @@ type EventData struct {
 	// TaskName is the name of the task that generated this event.
 	TaskName string
 
+	// Category is the category of the alert that generated this event.
+	Category string
+
 	// Concatenation of all group-by tags of the form [key=value,]+.
 	// If not groupBy is performed equal to literal 'nil'
 	Group string