+++ log.go
@@ -74,3 +74,4 @@ func (n *LogNode) Barrier(b edge.BarrierMessage) (edge.Message, error) {
 func (n *LogNode) DeleteGroup(d edge.DeleteGroupMessage) (edge.Message, error) {
 	return d, nil
 }
+func (n *LogNode) Done() {}